PMM adapter/main merge

+++ contracts/8/interfaces/IBalancer.sol
@@ -3,23 +3,21 @@ pragma solidity ^0.8.0;
 pragma abicoder v2;
 
 interface IBalancer {
-
     function getCurrentTokens() external view returns (address[] memory tokens);
 
     function swapExactAmountIn(
-        address tokenIn, 
-        uint tokenAmountIn, 
-        address tokenOut,  
-        uint minAmountOut,  
-        uint maxPrice
-    ) external returns (uint tokenAmountOut, uint spotPriceAfter);
+        address tokenIn,
+        uint256 tokenAmountIn,
+        address tokenOut,
+        uint256 minAmountOut,
+        uint256 maxPrice
+    ) external returns (uint256 tokenAmountOut, uint256 spotPriceAfter);
 
     function swapExactAmountOut(
         address tokenIn,
-        uint maxAmountIn,
+        uint256 maxAmountIn,
         address tokenOut,
-        uint tokenAmountOut,
-        uint maxPrice
-    ) external returns (uint tokenAmountIn, uint spotPriceAfter);
-
+        uint256 tokenAmountOut,
+        uint256 maxPrice
+    ) external returns (uint256 tokenAmountIn, uint256 spotPriceAfter);
 }
