+++ contracts/8/libraries/ZeroCopySource.sol
@@ -3,21 +3,28 @@ pragma solidity ^0.8.0;
 
 library ZeroCopySource {
     /* @notice              Read next byte as boolean type starting at offset from buff
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the boolean value
-    *  @return              The the read boolean value and new offset
-    */
-    function NextBool(bytes memory buff, uint256 offset) internal pure returns (bool, uint256) {
-        require(offset + 1 <= buff.length && offset < offset + 1, "Offset exceeds limit");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the boolean value
+     *  @return              The the read boolean value and new offset
+     */
+    function NextBool(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (bool, uint256)
+    {
+        require(
+            offset + 1 <= buff.length && offset < offset + 1,
+            "Offset exceeds limit"
+        );
         // byte === bytes1
         bytes1 v;
-        assembly{
+        assembly {
             v := mload(add(add(buff, 0x20), offset))
         }
         bool value;
         if (v == 0x01) {
-		    value = true;
-    	} else if (v == 0x00) {
+            value = true;
+        } else if (v == 0x00) {
             value = false;
         } else {
             revert("NextBool value error");
@@ -26,28 +33,42 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next byte starting at offset from buff
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the byte value
-    *  @return              The read byte value and new offset
-    */
-    function NextByte(bytes memory buff, uint256 offset) internal pure returns (bytes1, uint256) {
-        require(offset + 1 <= buff.length && offset < offset + 1, "NextByte, Offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the byte value
+     *  @return              The read byte value and new offset
+     */
+    function NextByte(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (bytes1, uint256)
+    {
+        require(
+            offset + 1 <= buff.length && offset < offset + 1,
+            "NextByte, Offset exceeds maximum"
+        );
         bytes1 v;
-        assembly{
+        assembly {
             v := mload(add(add(buff, 0x20), offset))
         }
         return (v, offset + 1);
     }
 
     /* @notice              Read next byte as uint8 starting at offset from buff
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the byte value
-    *  @return              The read uint8 value and new offset
-    */
-    function NextUint8(bytes memory buff, uint256 offset) internal pure returns (uint8, uint256) {
-        require(offset + 1 <= buff.length && offset < offset + 1, "NextUint8, Offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the byte value
+     *  @return              The read uint8 value and new offset
+     */
+    function NextUint8(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint8, uint256)
+    {
+        require(
+            offset + 1 <= buff.length && offset < offset + 1,
+            "NextUint8, Offset exceeds maximum"
+        );
         uint8 v;
-        assembly{
+        assembly {
             let tmpbytes := mload(0x40)
             let bvalue := mload(add(add(buff, 0x20), offset))
             mstore8(tmpbytes, byte(0, bvalue))
@@ -58,13 +79,20 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next two bytes as uint16 type starting from offset
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the uint16 value
-    *  @return              The read uint16 value and updated offset
-    */
-    function NextUint16(bytes memory buff, uint256 offset) internal pure returns (uint16, uint256) {
-        require(offset + 2 <= buff.length && offset < offset + 2, "NextUint16, offset exceeds maximum");
-        
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the uint16 value
+     *  @return              The read uint16 value and updated offset
+     */
+    function NextUint16(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint16, uint256)
+    {
+        require(
+            offset + 2 <= buff.length && offset < offset + 2,
+            "NextUint16, offset exceeds maximum"
+        );
+
         uint16 v;
         assembly {
             let tmpbytes := mload(0x40)
@@ -78,12 +106,19 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next four bytes as uint32 type starting from offset
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the uint32 value
-    *  @return              The read uint32 value and updated offset
-    */
-    function NextUint32(bytes memory buff, uint256 offset) internal pure returns (uint32, uint256) {
-        require(offset + 4 <= buff.length && offset < offset + 4, "NextUint32, offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the uint32 value
+     *  @return              The read uint32 value and updated offset
+     */
+    function NextUint32(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint32, uint256)
+    {
+        require(
+            offset + 4 <= buff.length && offset < offset + 4,
+            "NextUint32, offset exceeds maximum"
+        );
         uint32 v;
         assembly {
             let tmpbytes := mload(0x40)
@@ -95,7 +130,7 @@ library ZeroCopySource {
             } lt(tindex, byteLen) {
                 tindex := add(tindex, 0x01)
                 bindex := sub(bindex, 0x01)
-            }{
+            } {
                 mstore8(add(tmpbytes, tindex), byte(bindex, bvalue))
             }
             mstore(0x40, add(tmpbytes, byteLen))
@@ -105,12 +140,19 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next eight bytes as uint64 type starting from offset
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the uint64 value
-    *  @return              The read uint64 value and updated offset
-    */
-    function NextUint64(bytes memory buff, uint256 offset) internal pure returns (uint64, uint256) {
-        require(offset + 8 <= buff.length && offset < offset + 8, "NextUint64, offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the uint64 value
+     *  @return              The read uint64 value and updated offset
+     */
+    function NextUint64(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint64, uint256)
+    {
+        require(
+            offset + 8 <= buff.length && offset < offset + 8,
+            "NextUint64, offset exceeds maximum"
+        );
         uint64 v;
         assembly {
             let tmpbytes := mload(0x40)
@@ -122,7 +164,7 @@ library ZeroCopySource {
             } lt(tindex, byteLen) {
                 tindex := add(tindex, 0x01)
                 bindex := sub(bindex, 0x01)
-            }{
+            } {
                 mstore8(add(tmpbytes, tindex), byte(bindex, bvalue))
             }
             mstore(0x40, add(tmpbytes, byteLen))
@@ -137,8 +179,15 @@ library ZeroCopySource {
     *  @param offset        The position from where we read the uint256 value
     *  @return              The read uint256 value and updated offset
     */
-    function NextUint255(bytes memory buff, uint256 offset) internal pure returns (uint256, uint256) {
-        require(offset + 32 <= buff.length && offset < offset + 32, "NextUint255, offset exceeds maximum");
+    function NextUint255(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint256, uint256)
+    {
+        require(
+            offset + 32 <= buff.length && offset < offset + 32,
+            "NextUint255, offset exceeds maximum"
+        );
         uint256 v;
         assembly {
             let tmpbytes := mload(0x40)
@@ -150,13 +199,17 @@ library ZeroCopySource {
             } lt(tindex, byteLen) {
                 tindex := add(tindex, 0x01)
                 bindex := sub(bindex, 0x01)
-            }{
+            } {
                 mstore8(add(tmpbytes, tindex), byte(bindex, bvalue))
             }
             mstore(0x40, add(tmpbytes, byteLen))
             v := mload(tmpbytes)
         }
-        require(v <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, "Value exceeds the range");
+        require(
+            v <=
+                0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff,
+            "Value exceeds the range"
+        );
         return (v, offset + 32);
     }
 
@@ -166,12 +219,19 @@ library ZeroCopySource {
     *  @param offset        The position from where we read the bytes value
     *  @return              The read variable bytes array value and updated offset
     */
-    function NextVarBytes(bytes memory buff, uint256 offset) internal pure returns(bytes memory, uint256) {
-        uint len;
+    function NextVarBytes(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (bytes memory, uint256)
+    {
+        uint256 len;
         (len, offset) = NextVarUint(buff, offset);
-        require(offset + len <= buff.length && offset < offset + len, "NextVarBytes, offset exceeds maximum");
+        require(
+            offset + len <= buff.length && offset < offset + len,
+            "NextVarBytes, offset exceeds maximum"
+        );
         bytes memory tempBytes;
-        assembly{
+        assembly {
             switch iszero(len)
             case 0 {
                 // Get a location of some free memory and store it in tempBytes as
@@ -192,13 +252,19 @@ library ZeroCopySource {
                 // because when slicing multiples of 32 bytes (lengthmod == 0)
                 // the following copy loop was copying the origin's length
                 // and then ending prematurely not copying everything it should.
-                let mc := add(add(tempBytes, lengthmod), mul(0x20, iszero(lengthmod)))
+                let mc := add(
+                    add(tempBytes, lengthmod),
+                    mul(0x20, iszero(lengthmod))
+                )
                 let end := add(mc, len)
 
                 for {
                     // The multiplication in the next line has the same exact purpose
                     // as the one above.
-                    let cc := add(add(add(buff, lengthmod), mul(0x20, iszero(lengthmod))), offset)
+                    let cc := add(
+                        add(add(buff, lengthmod), mul(0x20, iszero(lengthmod))),
+                        offset
+                    )
                 } lt(mc, end) {
                     mc := add(mc, 0x20)
                     cc := add(cc, 0x20)
@@ -224,12 +290,19 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next 32 bytes starting from offset,
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the bytes value
-    *  @return              The read bytes32 value and updated offset
-    */
-    function NextHash(bytes memory buff, uint256 offset) internal pure returns (bytes32 , uint256) {
-        require(offset + 32 <= buff.length && offset < offset + 32, "NextHash, offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the bytes value
+     *  @return              The read bytes32 value and updated offset
+     */
+    function NextHash(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (bytes32, uint256)
+    {
+        require(
+            offset + 32 <= buff.length && offset < offset + 32,
+            "NextHash, offset exceeds maximum"
+        );
         bytes32 v;
         assembly {
             v := mload(add(buff, add(offset, 0x20)))
@@ -238,12 +311,19 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next 20 bytes starting from offset,
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the bytes value
-    *  @return              The read bytes20 value and updated offset
-    */
-    function NextBytes20(bytes memory buff, uint256 offset) internal pure returns (bytes20 , uint256) {
-        require(offset + 20 <= buff.length && offset < offset + 20, "NextBytes20, offset exceeds maximum");
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the bytes value
+     *  @return              The read bytes20 value and updated offset
+     */
+    function NextBytes20(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (bytes20, uint256)
+    {
+        require(
+            offset + 20 <= buff.length && offset < offset + 20,
+            "NextBytes20, offset exceeds maximum"
+        );
         bytes20 v;
         assembly {
             v := mload(add(buff, add(offset, 0x20)))
@@ -252,31 +332,41 @@ library ZeroCopySource {
     }
 
     /* @notice              Read next uint starting from offset,
-    *  @param buff          Source bytes array
-    *  @param offset        The position from where we read the uint value
-    *  @return              The read uint value and updated offset
-    */  
-    function NextVarUint(bytes memory buff, uint256 offset) internal pure returns(uint, uint256) {
+     *  @param buff          Source bytes array
+     *  @param offset        The position from where we read the uint value
+     *  @return              The read uint value and updated offset
+     */
+    function NextVarUint(bytes memory buff, uint256 offset)
+        internal
+        pure
+        returns (uint256, uint256)
+    {
         bytes1 v;
         (v, offset) = NextByte(buff, offset);
 
-        uint value;
+        uint256 value;
         if (v == 0xFD) {
             // return NextUint16(buff, offset);
             (value, offset) = NextUint16(buff, offset);
-            require(value >= 0xFD && value <= 0xFFFF, "NextUint16, value outside range");
+            require(
+                value >= 0xFD && value <= 0xFFFF,
+                "NextUint16, value outside range"
+            );
             return (value, offset);
         } else if (v == 0xFE) {
             // return NextUint32(buff, offset);
             (value, offset) = NextUint32(buff, offset);
-            require(value > 0xFFFF && value <= 0xFFFFFFFF, "NextVarUint, value outside range");
+            require(
+                value > 0xFFFF && value <= 0xFFFFFFFF,
+                "NextVarUint, value outside range"
+            );
             return (value, offset);
         } else if (v == 0xFF) {
             // return NextUint64(buff, offset);
             (value, offset) = NextUint64(buff, offset);
             require(value > 0xFFFFFFFF, "NextVarUint, value outside range");
             return (value, offset);
-        } else{
+        } else {
             // return (uint8(v), offset);
             value = uint8(v);
             require(value < 0xFD, "NextVarUint, value outside range");
