+++ contracts/8/test/SerializeTest.sol
@@ -8,12 +8,15 @@ import "../libraries/ZeroCopySink.sol";
 import "../libraries/ZeroCopySource.sol";
 
 library Utils {
-
     /* @notice      Convert the bytes array to bytes32 type, the bytes array length must be 32
-    *  @param _bs   Source bytes array
-    *  @return      bytes32
-    */
-    function bytesToBytes32(bytes memory _bs) internal pure returns (bytes32 value) {
+     *  @param _bs   Source bytes array
+     *  @return      bytes32
+     */
+    function bytesToBytes32(bytes memory _bs)
+        internal
+        pure
+        returns (bytes32 value)
+    {
         require(_bs.length == 32, "bytes length is not 32.");
         assembly {
             // load 32 bytes from memory starting from position _bs + 0x20 since the first 0x20 bytes stores _bs length
@@ -21,25 +24,41 @@ library Utils {
         }
     }
 
-    /* @notice      Convert bytes to uint256
-    *  @param _b    Source bytes should have length of 32
-    *  @return      uint256
-    */
-    function bytesToUint256(bytes memory _bs) internal pure returns (uint256 value) {
+    /*  @notice      Convert bytes to uint256
+     *  @param _b    Source bytes should have length of 32
+     *  @return      uint256
+     */
+    function bytesToUint256(bytes memory _bs)
+        internal
+        pure
+        returns (uint256 value)
+    {
         require(_bs.length == 32, "bytes length is not 32.");
         assembly {
             // load 32 bytes from memory starting from position _bs + 32
             value := mload(add(_bs, 0x20))
         }
-        require(value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, "Value exceeds the range");
+        require(
+            value <=
+                0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff,
+            "Value exceeds the range"
+        );
     }
 
-    /* @notice      Convert uint256 to bytes
-    *  @param _b    uint256 that needs to be converted
-    *  @return      bytes
-    */
-    function uint256ToBytes(uint256 _value) internal pure returns (bytes memory bs) {
-        require(_value <= 0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, "Value exceeds the range");
+    /*  @notice      Convert uint256 to bytes
+     *  @param _b    uint256 that needs to be converted
+     *  @return      bytes
+     */
+    function uint256ToBytes(uint256 _value)
+        internal
+        pure
+        returns (bytes memory bs)
+    {
+        require(
+            _value <=
+                0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff,
+            "Value exceeds the range"
+        );
         assembly {
             // Get a location of some free memory and store it in result as
             // Solidity does for memory variables.
@@ -54,10 +73,13 @@ library Utils {
     }
 
     /* @notice      Convert bytes to address
-    *  @param _bs   Source bytes: bytes length must be 20
-    *  @return      Converted address from source bytes
-    */
-    function bytesToAddress(bytes memory _bs) internal pure returns (address addr)
+     *  @param _bs   Source bytes: bytes length must be 20
+     *  @return      Converted address from source bytes
+     */
+    function bytesToAddress(bytes memory _bs)
+        internal
+        pure
+        returns (address addr)
     {
         require(_bs.length == 20, "bytes length does not match address");
         assembly {
@@ -65,14 +87,17 @@ library Utils {
             // load 32 bytes from mem[_bs+20], convert it into Uint160, meaning we take last 20 bytes as addr (address).
             addr := mload(add(_bs, 0x14))
         }
-
     }
-    
+
     /* @notice      Convert address to bytes
-    *  @param _addr Address need to be converted
-    *  @return      Converted bytes from address
-    */
-    function addressToBytes(address _addr) internal pure returns (bytes memory bs){
+     *  @param _addr Address need to be converted
+     *  @return      Converted bytes from address
+     */
+    function addressToBytes(address _addr)
+        internal
+        pure
+        returns (bytes memory bs)
+    {
         assembly {
             // Get a location of some free memory and store it in result as
             // Solidity does for memory variables.
@@ -83,23 +108,31 @@ library Utils {
             mstore(add(bs, 0x20), shl(96, _addr))
             // Update the free-memory pointer by padding our last write location to 32 bytes
             mstore(0x40, add(bs, 0x40))
-       }
+        }
     }
 
     /* @notice          Do hash leaf as the multi-chain does
-    *  @param _data     Data in bytes format
-    *  @return          Hashed value in bytes32 format
-    */
-    function hashLeaf(bytes memory _data) internal pure returns (bytes32 result)  {
+     *  @param _data     Data in bytes format
+     *  @return          Hashed value in bytes32 format
+     */
+    function hashLeaf(bytes memory _data)
+        internal
+        pure
+        returns (bytes32 result)
+    {
         result = sha256(abi.encodePacked(bytes1(0x0), _data));
     }
 
     /* @notice          Do hash children as the multi-chain does
-    *  @param _l        Left node
-    *  @param _r        Right node
-    *  @return          Hashed value in bytes32 format
-    */
-    function hashChildren(bytes32 _l, bytes32  _r) internal pure returns (bytes32 result)  {
+     *  @param _l        Left node
+     *  @param _r        Right node
+     *  @return          Hashed value in bytes32 format
+     */
+    function hashChildren(bytes32 _l, bytes32 _r)
+        internal
+        pure
+        returns (bytes32 result)
+    {
         result = sha256(abi.encodePacked(bytes1(0x01), _l, _r));
     }
 
@@ -112,13 +145,9 @@ library Utils {
     */
     function slice(
         bytes memory _bytes,
-        uint _start,
-        uint _length
-    )
-        internal
-        pure
-        returns (bytes memory)
-    {
+        uint256 _start,
+        uint256 _length
+    ) internal pure returns (bytes memory) {
         require(_bytes.length >= (_start + _length));
 
         bytes memory tempBytes;
@@ -145,13 +174,22 @@ library Utils {
                 // because when slicing multiples of 32 bytes (lengthmod == 0)
                 // the following copy loop was copying the origin's length
                 // and then ending prematurely not copying everything it should.
-                let mc := add(add(tempBytes, lengthmod), mul(0x20, iszero(lengthmod)))
+                let mc := add(
+                    add(tempBytes, lengthmod),
+                    mul(0x20, iszero(lengthmod))
+                )
                 let end := add(mc, _length)
 
                 for {
                     // The multiplication in the next line has the same exact purpose
                     // as the one above.
-                    let cc := add(add(add(_bytes, lengthmod), mul(0x20, iszero(lengthmod))), _start)
+                    let cc := add(
+                        add(
+                            add(_bytes, lengthmod),
+                            mul(0x20, iszero(lengthmod))
+                        ),
+                        _start
+                    )
                 } lt(mc, end) {
                     mc := add(mc, 0x20)
                     cc := add(cc, 0x20)
@@ -175,16 +213,21 @@ library Utils {
 
         return tempBytes;
     }
-    /* @notice              Check if the elements number of _signers within _keepers array is no less than _m
-    *  @param _keepers      The array consists of serveral address
-    *  @param _signers      Some specific addresses to be looked into
-    *  @param _m            The number requirement paramter
-    *  @return              True means containment, false meansdo do not contain.
-    */
-    function containMAddresses(address[] memory _keepers, address[] memory _signers, uint _m) internal pure returns (bool){
-        uint m = 0;
-        for(uint i = 0; i < _signers.length; i++){
-            for (uint j = 0; j < _keepers.length; j++) {
+
+    /*  @notice              Check if the elements number of _signers within _keepers array is no less than _m
+     *  @param _keepers      The array consists of serveral address
+     *  @param _signers      Some specific addresses to be looked into
+     *  @param _m            The number requirement paramter
+     *  @return              True means containment, false meansdo do not contain.
+     */
+    function containMAddresses(
+        address[] memory _keepers,
+        address[] memory _signers,
+        uint256 _m
+    ) internal pure returns (bool) {
+        uint256 m = 0;
+        for (uint256 i = 0; i < _signers.length; i++) {
+            for (uint256 j = 0; j < _keepers.length; j++) {
                 if (_signers[i] == _keepers[j]) {
                     m++;
                     delete _keepers[j];
@@ -194,21 +237,25 @@ library Utils {
         return m >= _m;
     }
 
-    /* @notice              TODO
-    *  @param key
-    *  @return
-    */
-    function compressMCPubKey(bytes memory key) internal pure returns (bytes memory newkey) {
-         require(key.length >= 67, "key lenggh is too short");
-         newkey = slice(key, 0, 35);
-         if (uint8(key[66]) % 2 == 0){
-             newkey[2] = bytes1(0x02);
-         } else {
-             newkey[2] = bytes1(0x03);
-         }
-         return newkey;
+    /*  @notice
+     *  @param key
+     *  @return
+     */
+    function compressMCPubKey(bytes memory key)
+        internal
+        pure
+        returns (bytes memory newkey)
+    {
+        require(key.length >= 67, "key lenggh is too short");
+        newkey = slice(key, 0, 35);
+        if (uint8(key[66]) % 2 == 0) {
+            newkey[2] = bytes1(0x02);
+        } else {
+            newkey[2] = bytes1(0x03);
+        }
+        return newkey;
     }
-    
+
     /**
      * @dev Returns true if `account` is a contract.
      *      Refer from https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/utils/Address.sol#L18
@@ -232,7 +279,9 @@ library Utils {
         bytes32 codehash;
         bytes32 accountHash = 0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470;
         // solhint-disable-next-line no-inline-assembly
-        assembly { codehash := extcodehash(account) }
+        assembly {
+            codehash := extcodehash(account)
+        }
         return (codehash != 0x0 && codehash != accountHash);
     }
 }
@@ -242,7 +291,6 @@ library Utils {
 /// @notice Explain to an end user what this does
 /// @dev Explain to a developer any extra details
 contract SerializeTest is OwnableUpgradeable {
-
     function initialize() public initializer {
         __Ownable_init();
     }
@@ -268,36 +316,52 @@ contract SerializeTest is OwnableUpgradeable {
     //-------------------------------
 
     /* @notice               Serialize Poly chain book keepers' info in Ethereum addresses format into raw bytes
-    *  @param keepersBytes   The serialized addresses
-    *  @return               serialized bytes result
-    */
-    function serializeAddress(address[] memory keepers) external pure returns (bytes memory) {
+     *  @param keepersBytes   The serialized addresses
+     *  @return               serialized bytes result
+     */
+    function serializeAddress(address[] memory keepers)
+        external
+        pure
+        returns (bytes memory)
+    {
         uint256 keeperLen = keepers.length;
         bytes memory keepersBytes = ZeroCopySink.WriteUint64(uint64(keeperLen));
-        for(uint i = 0; i < keeperLen; i++) {
-            keepersBytes = abi.encodePacked(keepersBytes, ZeroCopySink.WriteVarBytes(Utils.addressToBytes(keepers[i])));
+        for (uint256 i = 0; i < keeperLen; i++) {
+            keepersBytes = abi.encodePacked(
+                keepersBytes,
+                ZeroCopySink.WriteVarBytes(Utils.addressToBytes(keepers[i]))
+            );
         }
         return keepersBytes;
     }
 
     /* @notice               Deserialize bytes into Ethereum addresses
-    *  @param keepersBytes   The serialized addresses derived from Poly chain book keepers in bytes format
-    *  @return               addresses
-    */
-    function deserializeAddress(bytes memory keepersBytes) external pure returns (address[] memory) {
+     *  @param keepersBytes   The serialized addresses derived from Poly chain book keepers in bytes format
+     *  @return               addresses
+     */
+    function deserializeAddress(bytes memory keepersBytes)
+        external
+        pure
+        returns (address[] memory)
+    {
         uint256 off = 0;
         uint64 keeperLen;
         (keeperLen, off) = ZeroCopySource.NextUint64(keepersBytes, off);
         address[] memory keepers = new address[](keeperLen);
         bytes memory keeperBytes;
-        for(uint i = 0; i < keeperLen; i++) {
+        for (uint256 i = 0; i < keeperLen; i++) {
             (keeperBytes, off) = ZeroCopySource.NextVarBytes(keepersBytes, off);
             keepers[i] = Utils.bytesToAddress(keeperBytes);
         }
         return keepers;
     }
 
-    function crossChain(uint64 toChainId, bytes calldata toContract, bytes calldata method, bytes calldata txData) external view returns (bytes memory) {
+    function crossChain(
+        uint64 toChainId,
+        bytes calldata toContract,
+        bytes calldata method,
+        bytes calldata txData
+    ) external view returns (bytes memory) {
         // To help differentiate two txs, the ethTxHashIndex is increasing automatically
         uint256 txHashIndex = 0;
 
@@ -307,7 +371,11 @@ contract SerializeTest is OwnableUpgradeable {
         // Construct the makeTxParam, and put the hash info storage, to help provide proof of tx existence
         bytes memory rawParam = abi.encodePacked(
             ZeroCopySink.WriteVarBytes(paramTxHash),
-            ZeroCopySink.WriteVarBytes(abi.encodePacked(sha256(abi.encodePacked(address(this), paramTxHash)))),
+            ZeroCopySink.WriteVarBytes(
+                abi.encodePacked(
+                    sha256(abi.encodePacked(address(this), paramTxHash))
+                )
+            ),
             ZeroCopySink.WriteVarBytes(Utils.addressToBytes(msg.sender)),
             ZeroCopySink.WriteUint64(toChainId),
             ZeroCopySink.WriteVarBytes(toContract),
@@ -336,8 +404,16 @@ contract SerializeTest is OwnableUpgradeable {
         return rawParam;
     }
 
-    function deserializeBaseParam(bytes memory paramsBytes) external pure returns (
-        address fromToken, address toToken, uint256 amount, uint256 minAmount, uint256 deadLine)
+    function deserializeBaseParam(bytes memory paramsBytes)
+        external
+        pure
+        returns (
+            address fromToken,
+            address toToken,
+            uint256 amount,
+            uint256 minAmount,
+            uint256 deadLine
+        )
     {
         uint256 off = 0;
         bytes memory paramBytes;
